[Core] Use pipeline Agent as canonical
